[Fix] #54 - CourseDiscoveryVC UI 컬렉션뷰로 수정
