--- Runnect-iOS/Runnect-iOS/Presentation/CourseDiscovery/Reuse/CollectionViewCell.swift
@@ -1,114 +1,0 @@
-//
-//  CollectionViewCell.swift
-//  Runnect-iOS
-//
-//  Created by YEONOO on 2023/01/10.
-//
-
-import UIKit
-import SnapKit
-
-import Then
-
-class CollectionViewCell: UICollectionViewCell {
-    
-    // MARK: - Identifier
-    static let identifier = "CollectionViewCell"
-    // MARK: - collectionview
-    private lazy var mapCollectionView: UICollectionView = {
-        let layout = UICollectionViewFlowLayout()
-        layout.scrollDirection = .vertical
-
-        let collectionView = UICollectionView(frame: .zero, collectionViewLayout: layout)
-        collectionView.backgroundColor = .clear
-        collectionView.translatesAutoresizingMaskIntoConstraints = false
-        collectionView.isScrollEnabled = false
-        collectionView.showsVerticalScrollIndicator = false
-        return collectionView
-    }()
-    // MARK: - Constants
-    
-    final let collectionViewInset = UIEdgeInsets(top: 28, left: 16, bottom: 28, right: 16)
-    final let itemSpacing: CGFloat = 10
-    final let lineSpacing: CGFloat = 20
-   
-    // MARK: - Life cycle
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        layout()
-        register()
-        setDelegate()
-        layout()
-    }
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
-
-extension CollectionViewCell {
-    
-    private func setDelegate() {
-        mapCollectionView.delegate = self
-        mapCollectionView.dataSource = self
-    }
-    private func register() {
-        mapCollectionView.register(CourseListCVC.self,
-                                          forCellWithReuseIdentifier: CourseListCVC.className)
-    }
-}
-// MARK: - Extensions
-
-extension CollectionViewCell {
-    
-    // MARK: - Layout Helpers
-    
-    func layout() {
-        contentView.backgroundColor = .clear
-        contentView.addSubview(mapCollectionView)
-        mapCollectionView.snp.makeConstraints {
-            $0.top.equalToSuperview()
-            $0.leading.trailing.equalTo(contentView.safeAreaLayoutGuide)
-            $0.bottom.equalToSuperview()
-            $0.height.equalTo(1000)
-        }
-    }
-}
-// MARK: - UICollectionViewDelegate, UICollectionViewDataSource
-
-extension CollectionViewCell: UICollectionViewDelegate, UICollectionViewDataSource {
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return 15
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: CourseListCVC.className,
-                                                            for: indexPath)
-                as? CourseListCVC else { return UICollectionViewCell() }
-        cell.setCellType(type: .all)
-        return cell
-    }
-}
-
-// MARK: - UICollectionViewDelegateFlowLayout
-
-extension CollectionViewCell: UICollectionViewDelegateFlowLayout {
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        let cellWidth = (UIScreen.main.bounds.width - (self.itemSpacing + 2*self.collectionViewInset.left)) / 2
-        let cellHeight = CourseListCVCType.getCellHeight(type: .all, cellWidth: cellWidth)
-        
-        return CGSize(width: cellWidth, height: cellHeight)
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, insetForSectionAt section: Int) -> UIEdgeInsets {
-        return self.collectionViewInset
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumInteritemSpacingForSectionAt section: Int) -> CGFloat {
-        return self.itemSpacing
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumLineSpacingForSectionAt section: Int) -> CGFloat {
-        return self.lineSpacing
-    }
-
-}
