+++ Runnect-iOS/Runnect-iOS/Presentation/CourseDiscovery/Reuse/MapCollectionViewCell.swift
@@ -1,8 +1,8 @@
 //
-//  MapCollectionViewCell.swift
+//  CollectionViewCell.swift
 //  Runnect-iOS
 //
-//  Created by YEONOO on 2023/01/02.
+//  Created by YEONOO on 2023/01/10.
 //
 
 import UIKit
@@ -13,110 +13,102 @@ import Then
 class MapCollectionViewCell: UICollectionViewCell {
     
     // MARK: - Identifier
-    static let identifier = "MapCollectionViewCell"
-    // MARK: - UI Components
-    private let mapContainerView = UIView()
-    private let mapImageView = UIImageView().then {
-        $0.image = ImageLiterals.imgLogo
-        $0.layer.cornerRadius = 5
-    }
-    private let titleLabel = UILabel().then {
-        $0.text = "제목제목제목제목제목"
-        $0.font =  UIFont.b4
-        $0.textColor = UIColor.black
-    }
-    private let locationLabel = UILabel().then {
-        $0.text = "00시 00구"
-        $0.font = UIFont.b6
-        $0.textColor = UIColor.g2
-    }
-    let heartButton = UIImageView().then {
-        $0.image = ImageLiterals.icHeartFill
-    }
-    // MARK: - ClickAction Constants
-    var clickCount: Int = 0 {
-            didSet {
-                if clickCount == 0 {
-                    mapImageView.backgroundColor = UIColor.systemGray4
-                    mapImageView.alpha = 1
-                    mapImageView.layer.borderColor = UIColor.w1.cgColor
-                    mapImageView.layer.borderWidth = 0 } else {
-                    mapImageView.backgroundColor = UIColor.m1
-                    mapImageView.alpha = 0.15
-                    mapImageView.layer.borderColor = UIColor.purple.cgColor
-                    mapImageView.layer.borderWidth = 2
-                }
-            }
-        }
-        override var isSelected: Bool {
-            didSet {
-                if !isSelected {
-                    mapImageView.backgroundColor = UIColor.systemGray4
-                    mapImageView.alpha = 1
-                    mapImageView.layer.borderColor = UIColor.w1.cgColor
-                    mapImageView.layer.borderWidth = 0
-                    clickCount = 0
-                }
-            }
-        }
+    static let collectionViewId = "CollectionViewCell"
+    // MARK: - collectionview
+    private lazy var mapCollectionView: UICollectionView = {
+        let layout = UICollectionViewFlowLayout()
+        layout.scrollDirection = .vertical
+
+        let collectionView = UICollectionView(frame: .zero, collectionViewLayout: layout)
+        collectionView.backgroundColor = .clear
+        collectionView.translatesAutoresizingMaskIntoConstraints = false
+        collectionView.isScrollEnabled = false
+        collectionView.showsVerticalScrollIndicator = false
+        return collectionView
+    }()
+    // MARK: - Constants
+    
+    final let collectionViewInset = UIEdgeInsets(top: 28, left: 16, bottom: 28, right: 16)
+    final let itemSpacing: CGFloat = 10
+    final let lineSpacing: CGFloat = 20
+   
     // MARK: - Life cycle
     override init(frame: CGRect) {
         super.init(frame: frame)
         layout()
-        
+        register()
+        setDelegate()
+        layout()
     }
-    
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
-    
 }
 
+extension MapCollectionViewCell {
+    
+    private func setDelegate() {
+        mapCollectionView.delegate = self
+        mapCollectionView.dataSource = self
+    }
+    private func register() {
+        mapCollectionView.register(CourseListCVC.self,
+                                          forCellWithReuseIdentifier: CourseListCVC.className)
+    }
+}
 // MARK: - Extensions
 
 extension MapCollectionViewCell {
-    // MARK: - Function
-   
+    
     // MARK: - Layout Helpers
     
     func layout() {
         contentView.backgroundColor = .clear
-        mapImageView.backgroundColor = .systemGray4
-        [mapContainerView, mapImageView, titleLabel, locationLabel, heartButton].forEach {
-            contentView.addSubview($0)}
-        mapContainerView.addSubview(mapImageView)
-        mapContainerView.snp.makeConstraints {
+        contentView.addSubview(mapCollectionView)
+        mapCollectionView.snp.makeConstraints {
             $0.top.equalToSuperview()
-            $0.centerX.equalToSuperview()
-            $0.leading.equalToSuperview()
-            $0.height.equalTo(161)
-        }
-        mapImageView.snp.makeConstraints {
-            $0.top.leading.trailing.equalToSuperview()
-            $0.height.equalTo(110)
-            
-        }
-        titleLabel.snp.makeConstraints {
-            $0.top.equalTo(mapImageView.snp.bottom).offset(4)
-            $0.leading.equalTo(self.mapContainerView.snp.leading)
-        }
-        heartButton.snp.makeConstraints {
-            $0.top.equalTo(mapImageView.snp.bottom).offset(4)
-            $0.width.equalTo(14)
-            $0.height.equalTo(12)
-            $0.trailing.equalTo(self.mapContainerView.snp.trailing)
-        }
-        locationLabel.snp.makeConstraints {
-            $0.top.equalTo(titleLabel.snp.bottom).offset(1)
-            $0.leading.equalTo(self.mapContainerView.snp.leading)
+            $0.leading.trailing.equalTo(contentView.safeAreaLayoutGuide)
+            $0.bottom.equalToSuperview()
+//            $0.height.equalTo(1000)
         }
     }
-    
-    // MARK: - General Helpers
-    
-    func dataBind(model: MapModel) {
-        mapImageView.image = UIImage(named: model.mapImage)
-        titleLabel.text = model.title
-        locationLabel.text = model.location
+}
+// MARK: - UICollectionViewDelegate, UICollectionViewDataSource
+
+extension MapCollectionViewCell: UICollectionViewDelegate, UICollectionViewDataSource {
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        return 15
     }
+    
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: CourseListCVC.className,
+                                                            for: indexPath)
+                as? CourseListCVC else { return UICollectionViewCell() }
+        cell.setCellType(type: .all)
+        return cell
+    }
+}
+
+// MARK: - UICollectionViewDelegateFlowLayout
+
+extension MapCollectionViewCell: UICollectionViewDelegateFlowLayout {
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
+        let cellWidth = (UIScreen.main.bounds.width - (self.itemSpacing + 2*self.collectionViewInset.left)) / 2
+        let cellHeight = CourseListCVCType.getCellHeight(type: .all, cellWidth: cellWidth)
+        
+        return CGSize(width: cellWidth, height: cellHeight)
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, insetForSectionAt section: Int) -> UIEdgeInsets {
+        return self.collectionViewInset
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumInteritemSpacingForSectionAt section: Int) -> CGFloat {
+        return self.itemSpacing
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumLineSpacingForSectionAt section: Int) -> CGFloat {
+        return self.lineSpacing
+    }
+
 }
